Add the Game of Life example, ported from Bevy
The shader for this example is entirely unchanged. The Rust side of the
example is half as long as the Bevy version, and conceptually FAR
simpler.

This did require a little new functionality, because their Life example
uses double buffering. It has two texture buffers that it swaps every
frame. In order to enable this, the ShaderBufferSet has two new
functions:
- swap_buffers will swap the bindings of two buffers
- set_storage_texture_access changes whether a texture is Read, Write or
ReadWrite

This is not actually the best approach to solving this problem, as it
requires the buffers to be swapped in the main world without direct
knowledge of the state of the shaders. There should instead be explicit
support for double buffering, so it can be managed by the ComputeNode in
the render world. This will come in a future change.

This also required that the TextureFormat can be specified directly.
Previously, all textures were Rgba8Unorm, but the game of life example
used R32Float. Now textures can use any format. The default fill color
was also solid black before, but that can no longer be assumed when the
format could be anything, so now a fill value must be specified.

The initial access specified is also now passed in, since the Game of
Life example requires one of the textures to initially be Read and the
other initially Write. These are both created with add_write_texture,
which takes an argument for the access type. This is confusing, and will
be addressed in a future change.