Add much better support for double buffering
This removes the swap_buffers and set_storage_texture_access, which were
hacks to allow the Game of Life demo to do double buffering. Now,
instead, any buffer can be created with two bindings, and if so, it also
has two backing buffers. At any given time, one of them is considered
the front buffer, and gets the first binding, and the other gets the
second binding. If they're storage textures, then it ignores the
provided access mode. The front buffer is always ReadOnly and the back
buffer is always WriteOnly.

This also adds a new action to PipelineData, called SwapBuffers. This
allows precise control over when buffer swaps occur, but if you leave
this out, the buffers will never swap.