Rename all parts of the compute sequence
It used to be the ActiveComputePipeline, which contained a list of
PipelineGroups, which contained PipelineSteps, and PipelineData. Now,
it's a ComputeSequence, which contains a list of ComputeTasks, which
contains ComputeSteps, each of which has a ComputeAction.

No functional changes. Just better naming.

Aside from being more descriptive, this no longer collides with WGPU's
usage of the word pipeline in compute shaders.