Remove add_read_write_texture
The read write texture was a weird hack whose purpose is entirely
superceded by double buffers. In addition to just removing this
function:

- ComputeAction::CopyTexture was removed, as it existed entirely to copy
between the two textures that were created by read/write texture.
- ShaderBufferStorage::Texture was removed, as it was only used by the
read/write textures. The other textures used StorageTexture.
- add_write_texture was renamed to add_texture_fill. It can have any
access type now, so it no longer makes sense to label it as a write
texture. But it's initialized with a fill, so specifying that in the
name makes it consistent with the other buffer creation functons.