Update to Bevy 0.15
The only change to the public API is based on the change to the Sprite
component. It now an image field on it that stores the Handle<Image> to
display, instead of expecting a Handle<Image> as a separate component.
There's also required components now. So the DoubleBufferedSprite
component now requires a Sprite component, and the swap_sprite_buffers
system now looks for a Sprite instead of a Handle<Image>.